Refactor again * Rename ctxToContext to ctx-to-context * Change to ignore dist * Delete lib * Add prepublish task for npm publish
