+++ .gitignore
@@ -18,3 +18,4 @@ Thumbs.db
 .svn/
 *.swp
 out/
+dist
