+++ src/index.js
@@ -1,1 +1,13 @@
-module.exports = require('./router');
+import FalcorRouter from 'falcor-router';
+import dataSourceRoute from './falcor-koa';
+
+const router = {};
+
+router.routes = (routes) => {
+  const Router = FalcorRouter.createClass(routes);
+  return dataSourceRoute(() => {
+    return new Router();
+  });
+};
+
+export default router;
